+++ constants/builtInJigsawImages.ts
